🎨 Improve project folder structure
1. Added a .gitignore
2. Added a setup.py for the project
3. Created a password-shack directory where the source code lives
4. Made the entire project into a python package
