+++ password_shack/__init__.py
@@ -1,0 +1,3 @@
+"""
+Command line tool to create, store and retrieve passwords securely
+"""
