+++ password_shack/login.py
@@ -1,0 +1,7 @@
+"""
+Module that deals with logging in or creating a new account
+for a user that uses the password manager
+"""
+
+def login_menu():
+    pass
